requirements.txt
django-calendarium
django-celery
django-revision
Django~=4.2.5
Faker~=19.6.2
git+https://github.com/botswana-harvard/edc-appointment.git@develop#egg=edc-appointment
git+https://github.com/botswana-harvard/edc-base.git@develop#egg=edc-base
git+https://github.com/botswana-harvard/edc-calendar.git@develop#egg=edc-calendar
git+https://github.com/botswana-harvard/edc-constants.git@develop#egg=edc-constants
git+https://github.com/botswana-harvard/edc-device.git@develop#egg=edc-device
git+https://github.com/botswana-harvard/edc-facility.git@develop#egg=edc-facility
git+https://github.com/botswana-harvard/edc-identifier.git@develop#egg=edc-identifier
git+https://github.com/botswana-harvard/edc-search.git@develop#egg=edc-search
git+https://github.com/botswana-harvard/edc-timepoint.git@develop#egg=edc-timepoint
git+https://github.com/botswana-harvard/edc-visit-schedule.git@develop#egg=edc-visit-schedule
git+https://github.com/flourishbhp/flourish@develop#egg=flourish
python-dateutil~=2.8.2
requests
setuptools~=68.2.2
